requirements.txt
absl-py==2.1.0
altgraph==0.17.4
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.16
django-cors-headers==4.4.0
django-filter==24.3
django-rest-passwordreset==1.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
flatbuffers==24.3.7
fonttools==4.50.0
fqdn==1.5.1
gast==0.4.0
google-api-core==2.15.0
google-auth==2.26.2
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-cloud-vision==3.5.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.2
importlib-resources==6.3.2
ipykernel==6.29.3
ipython==8.12.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==8.6.1
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter-server==2.13.0
jupyter-server-terminals==0.5.3
jupyterlab==4.1.5
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.4
keras==2.13.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mistune==3.0.2
mysql-connector-python==8.2.0
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.3
nest-asyncio==1.6.0
notebook==7.1.2
notebook-shim==0.2.4
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
overrides==7.7.0
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pefile==2024.8.26
pickleshare==0.7.5
pillow==10.2.0
pkgutil-resolve-name==1.3.10
platformdirs==4.2.0
prometheus-client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.21.12
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pygments==2.17.2
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.3
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
requests-oauthlib==1.4.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
scipy==1.10.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.1
stack-data==0.6.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
traitlets==5.14.2
types-python-dateutil==2.9.0.20240316
typing-extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
werkzeug==3.0.1
wrapt==1.16.0
zipp==3.18.1


{
    "refresh": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0b2tlbl90eXBlIjoicmVmcmVzaCIsImV4cCI6NDg4NDAzNzk0MiwiaWF0IjoxNzMwNDM3OTQyLCJqdGkiOiJlMTFmNDZjOTY5OTc0ZmU0ODU5NTJmMjVkYzU3ZTA1ZCIsInVzZXJfaWQiOjJ9.gi-o0lkHwRWCbDKbhoe-ui4s5f0apRf89XzttmZzIwQ",
    "access": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0b2tlbl90eXBlIjoiYWNjZXNzIiwiZXhwIjoxNzYxOTczOTQyLCJpYXQiOjE3MzA0Mzc5NDIsImp0aSI6ImE3ZjA5NTZlM2ViMDQ4MGM5NzJiOGY3NDRkZTdkZTdmIiwidXNlcl9pZCI6Mn0.I6XK_OI1PRtuoKbmpZJnQuF3N3iWqfbndiBtnto3pn4"
}